Work on an experimental new lexer/parser design
This one's more directly inspired by Lisp. Sigils (like ==) can be
as long and include as many different characters as one likes.
This also implements a "StringCache", so it deduplicates all the instances
of the same string while also simplifying lexing AND adding new keywords,
whether from within the compiler or as an extension.